+++ flask_server/requirements-base.txt
@@ -1,4 +1,5 @@
 Authlib
 flask
 gunicorn
+python-decouple
 requests
